+++ CloudShell.Hosting/Shell/CoreShellViews.cs
@@ -7,4 +7,6 @@ public static class CoreShellViews
     public const string Extensions = "cloudshell.core.extensions";
 
     public const string Users = "cloudshell.core.users";
+
+    public const string Settings = "cloudshell.core.settings";
 }